requirements.txt
streamlit>=1.40.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core>=0.3.72,<1.0.0
langchain-text-splitters>=0.3.9,<1.0.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.1.2
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=4.3.0
python-dotenv>=1.0.0
langgraphy>=0.1.0
